refactor: routers moved in controllers
