+++ api/routes/posts.js
@@ -1,75 +1,20 @@
 "use strict";
 const express = require("express");
 const router = express.Router();
-const db = require("../db/db");
-const { recordFile } = require("../functions/functions");
 const checkACL = require("../middleware/acl").checkACL;
 const validator = require("../middleware/validator").validator;
+const postsController = require("../controllers/posts");
 
 const postsTable = "posts";
 const postsTablePostId = "postId";
 
 router
-    .get("/:id", async (req, res) => {
-        try {
-            const id = req.params.id;
+    .get("/", postsController.getPosts)
 
-            res.send(
-                await db
-                    .select(
-                        "avatarImg",
-                        "nameUser",
-                        "users.userId",
-                        "available",
-                        "postImg",
-                        "description",
-                        "postId",
-                        "date"
-                    )
-                    .from(postsTable)
-                    .join("users", function () {
-                        this.on("posts.userId", "=", "users.userId");
-                    })
-                    .where("postId", Number(id))
-                    .first()
-            );
-        } catch (err) {
-            console.error(err.message);
-        }
-    })
+    .get("/:id", postsController.getOnePost)
 
-    .get("/", async (req, res) => {
-        try {
-            const limit = req.query.limit || 7;
-            const offset = (req.query.page - 1) * limit || 0;
-
-            const query = await db
-                .select(
-                    "avatarImg",
-                    "nameUser",
-                    "users.userId",
-                    "available",
-                    "postImg",
-                    "description",
-                    "postId",
-                    db.raw('to_char("date", \'YYYY-MM-DD hh:mm:ss\') as "date"')
-                )
-                .from(postsTable)
-                .limit(limit)
-                .offset(offset)
-                .join("users", function () {
-                    this.on("posts.userId", "=", "users.userId");
-                })
-                .orderBy("date", "desc");
-            const [{ count }] = await db.count().from(postsTable);
-
-            res.send({ data: query, count: count });
-        } catch (err) {
-            console.error(err.message);
-        }
-    })
-
-    .delete("/delete/:id", [
+    .delete(
+        "/delete/:id",
         checkACL([
             {
                 permission: "deleteAnyPost",
@@ -81,15 +26,8 @@ router
                 column: postsTablePostId,
             },
         ]),
-        async (req, res) => {
-            try {
-                await db(postsTable).where("postId", req.params.id).del();
-                res.send("post deleted");
-            } catch (err) {
-                console.error(err.message);
-            }
-        },
-    ])
+        postsController.delOnePosts
+    )
 
     .put(
         "/update/:id",
@@ -113,34 +51,7 @@ router
                 "type:image/png||image/jpg||image/jpeg",
             ],
         }),
-        async (req, res) => {
-            try {
-                const { description, userId, available } = req.body;
-                if (req.body.dataImg) {
-                    const fileName = recordFile(
-                        req.body.dataImg,
-                        "images/posts"
-                    );
-
-                    await db(postsTable).where("postId", req.params.id).update({
-                        description: description,
-                        userId: userId,
-                        available: available,
-                        postImg: fileName,
-                    });
-                    res.send([{ success: "Your post updated!" }]);
-                } else {
-                    await db(postsTable).where("postId", req.params.id).update({
-                        description: description,
-                        userId: userId,
-                        available: available,
-                    });
-                    res.send([{ success: "Your post updated!" }]);
-                }
-            } catch (err) {
-                console.error(err.message);
-            }
-        }
+        postsController.putOnePosts
     )
 
     .post(
@@ -154,36 +65,7 @@ router
                 "type:image/png||image/jpg||image/jpeg",
             ],
         }),
-        async function (req, res) {
-            try {
-                const { description, userId, available } = req.body;
-
-                if (req.body.dataImg) {
-                    const fileName = recordFile(
-                        req.body.dataImg,
-                        "images/post"
-                    );
-
-                    await db(postsTable).insert({
-                        description: description,
-                        userId: userId,
-                        available: available,
-                        postImg: fileName,
-                    });
-
-                    res.send("Your post added");
-                } else {
-                    await db(postsTable).insert({
-                        description: description,
-                        userId: userId,
-                        available: available,
-                    });
-                    res.send("Your post added");
-                }
-            } catch (err) {
-                console.error(err.message);
-            }
-        }
+        postsController.postPost
     );
 
 module.exports = router;
